Minor bug fix (explained below, expand commit):
Scenario: when a node is removed from the LazyPQ, it's removed rom the set but not the pq. Then it is added back with a differnet key.
This requires that the valie of the key ion the pq be updated, and not use the old value. Hence it is important that the == operator used by the queue
understand that the NodeKeyPairs with the same coordinates but having a different keys are different. They also must be stored differently in the set.